requirements.txt
altgraph==0.17.3
asgiref==3.6.0
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.10.0
future==0.18.3
macholib==1.16.2
Pillow==9.4.0
platformdirs==3.1.1
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
virtualenv==20.21.0
